% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateMI.R
\name{calculateMI}
\alias{calculateMI}
\title{Calculate MI, raw or weighted}
\usage{
calculateMI(
  msa_matrix,
  weighted = TRUE,
  groups = NULL,
  weights = "equal",
  ncores = 2
)
}
\arguments{
\item{msa_matrix}{Matrix format of MSA, from readMSA}

\item{weighted}{either TRUE, for weighted MI , or FALSE, for raw MI}

\item{groups}{a dataframe with 'ID' and 'group'}

\item{weights}{either "equal" or a dataframe with 'group' and 'weight'}

\item{ncores}{Number of cores for parallelization}
}
\value{
A data.frame containing entropy, MI, APC, and MIp values
}
\description{
This function calculates the MI, APC, and MIp from an MSA loaded with readMSA.
If weighted is set to FALSE, then only msa_matrix is needed as input. If
weighted is set to TRUE and weights is set to "equal", then a "groups"
dataframe is required that contains an 'ID' column that matches the msa_matrix
and a 'group' column. Otherwise, a weights dataframe is required that
contains a 'group' column and a 'weight' column.
}
