% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateMI.R
\name{get_ent_unweighted}
\alias{get_ent_unweighted}
\title{Get Shannon entropy, raw}
\usage{
get_ent_unweighted(mat)
}
\arguments{
\item{mat}{Matrix format of MSA, from readMSA}
}
\value{
A table containing the AA frequencies from mat
}
\description{
This function calculates the Shannon entropy of a column or pair of columns
from an MSA
}
