% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getNetworkInput}
\alias{getNetworkInput}
\title{Generate input format for associationsubgraphs network visualization}
\usage{
getNetworkInput(mi_output)
}
\arguments{
\item{mi_output}{output from calculateMI}
}
\value{
A dataframe containing columns 'a', 'b', and 'strength' (MIp)
}
\description{
This function takes the output from calculateMI and generates a data.frame
that can be used as input for the associationsubgraphs package
}
