% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{pb2_meta}
\alias{pb2_meta}
\title{This is data to be included in my package}
\description{
This is data to be included in my package
}
\references{
\url{gisaid.com}
}
\author{
Sarah Arcos \email{sarah.e.arcos@gmail.com}
}
\keyword{data}
