% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readMSA}
\alias{readMSA}
\title{Read in an MSA in fasta format}
\usage{
readMSA(file_path)
}
\arguments{
\item{file_path}{path to fasta file}
}
\value{
The MSA formatted as a matrix
}
\description{
This function reads in an MSA in fasta format and generates an MSA matrix.
The MSA matrix can then be used as input for calculateMI
}
